move fixtures to conftest in utils
